feat: redesenha HomeScreen focada em eventos + corrige setState após dispose + TabBar sempre visível
- HomeScreen completamente redesenhada focada em GESTÃO DE EVENTOS
- Mostra análise preditiva de clima para próximos eventos (até 6 meses)
- Resumo visual com contadores de eventos por risco (crítico/atenção/seguro)
- Usa EventWeatherPredictionService para análises inteligentes
- UI moderna com gradientes e glassmorphism
- Botão para alternar tema claro/escuro no AppBar
- AutomaticKeepAliveClientMixin mantém estado ao trocar tabs

- Corrige erro setState() após dispose no DisastersScreen
- Adiciona verificação mounted antes de todos setState()

- TabBar movida para bottomNavigationBar (sempre visível)
- Não desaparece mais ao fazer scroll

- Backup da HomeScreen antiga salvo em home_screen_old_backup.dart
